+++ tests/testthat/test-api.R
@@ -11,23 +11,21 @@ if (run.integration.tests) {
     test_that("crunchUserAgent", {
         expect_true(grepl("rcrunch", crunchUserAgent()))
         expect_true(grepl("libcurl", crunchUserAgent()))
-        expect_that(try(crunchUserAgent("anotherpackage/3.1.4")), is_not_an_error())
+        expect_that(try(crunchUserAgent("anotherpackage/3.1.4")),
+            is_not_an_error())
         expect_true(grepl("anotherpackage", crunchUserAgent("anotherpackage")))
     })
     
     with(test.authentication, 
         test_that("API root can be fetched", {
-            expect_that(try(getAPIroot()), is_not_an_error())
-            urls <- getAPIroot()
-            expect_true(is.shojiObject(urls))
+            expect_true(is.shojiObject(getAPIroot()))
         }))
 
-    if (crunchAPIcanBeReached()) {
-        test_that("API calls throw an error if user is not authenticated", {
-            logout()
-            expect_error(getAPIroot(), "401")
-        })
-    }
+    test_that("API calls throw an error if user is not authenticated", {
+        logout()
+        expect_error(getAPIroot(), 
+            "You are not authenticated. Please `login\\(\\)` and try again.")
+    })
     
     test_that("Deprecated endpoints tell user to upgrade", {
         expect_error(crGET("http://httpbin.org/status/410"),